% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cloud_point}
\alias{cloud_point}
\title{Cloud Point of a Liquid}
\format{
A data frame with 19 rows and 2 columns:
\describe{
  \item{percentage_i8}{Percentage of I-8 in the liquid mixture (0--10).}
  \item{cloud_point}{Cloud point temperature of the liquid (degrees Celsius).}
}
}
\source{
N.R. Draper and R. Craig Smith (1998).
\emph{Applied Regression Analysis}, 3rd edition.
John Wiley & Sons, Exercise 4, p. 629.

Available in \pkg{robustbase} as \code{cloud}.
}
\usage{
cloud_point
}
\description{
Measurements on the cloud point of a liquid, which is a measure of the
degree of crystallization in a stock.  The dataset contains three leverage
points (baseline measurements at \code{percentage_i8 = 0}) that pull
poorly-specified models away from the underlying trend.
}
\details{
This dataset is used in De Carvalho et al. (2017), Section 6.3, to
illustrate robustness to leverage points.  The recommended width estimator
is \code{sigma_method = "s3"}.
}
\section{Outlier structure}{

Observations 1, 10 and 16 have \code{percentage_i8 = 0} and act as leverage
points because they correspond to a baseline condition that is far from the
centroid of the predictor space.
}

\examples{
data(cloud_point)

fit_ols  <- lm(cloud_point ~ percentage_i8, data = cloud_point)
fit_gkrr <- gkrr(cloud_point ~ percentage_i8, data = cloud_point,
                 sigma_method = "s3")

plot(cloud_point$percentage_i8, cloud_point$cloud_point,
     xlab = "Percentage I-8", ylab = "Cloud point (C)",
     main = "Cloud Point Data",
     pch = 19, col = "grey60")
abline(fit_ols,  col = "black",     lwd = 2, lty = 2)
abline(fit_gkrr, col = "firebrick", lwd = 2)
legend("topleft", c("OLS", "GKRR"), col = c("black","firebrick"),
       lty = c(2,1), lwd = 2, bty = "n")
}
\references{
De Carvalho, F.A.T., Lima Neto, E.A., Ferreira, M.R.P. (2017).
\doi{10.1016/j.neucom.2016.12.035}
}
\keyword{datasets}
