% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delivery}
\alias{delivery}
\title{Soft-Drink Delivery Time Data}
\format{
A data frame with 25 rows and 3 columns:
\describe{
  \item{n_products}{Number of products stocked (cases).}
  \item{distance}{Distance walked by the driver (feet).}
  \item{delivery_time}{Total service time at the stop (minutes).}
}
}
\source{
D.C. Montgomery and E.A. Peck (1992).
\emph{Introduction to Linear Regression Analysis}, 2nd edition.
John Wiley & Sons, Table B.7.

Available in \pkg{robustbase} as \code{delivery}.
}
\usage{
delivery
}
\description{
Data on the time required by a route driver to service vending machines at
25 stops.  Two predictors (number of products delivered and distance
travelled) are used to model delivery time.  Observation 9 is a bad leverage
point: it has both a very large number of products (30) and a very large
distance (1460 feet), pulling regression fits away from the main trend.
}
\details{
This dataset is used in De Carvalho et al. (2017), Section 6.4, to
illustrate robustness to leverage points in multiple regression.  The
recommended width estimator is \code{sigma_method = "s3"}.
}
\section{Outlier structure}{

Observation 9 (\code{n_products = 30}, \code{distance = 1460}) is a
leverage point that simultaneously deviates in the predictor space and
has an unusually large response, making it a bad leverage point.
}

\examples{
data(delivery)

fit_ols  <- lm(delivery_time ~ n_products + distance, data = delivery)
fit_gkrr <- gkrr(delivery_time ~ n_products + distance, data = delivery,
                 sigma_method = "s3")

# Compare coefficients
rbind(OLS  = coef(fit_ols),
      GKRR = coef(fit_gkrr))

# Diagnostic plot: kernel weights highlight observation 9
plot(fit_gkrr, which = 4, ask = FALSE)
}
\references{
De Carvalho, F.A.T., Lima Neto, E.A., Ferreira, M.R.P. (2017).
\doi{10.1016/j.neucom.2016.12.035}
}
\keyword{datasets}
