% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gkrreg-package.R
\docType{package}
\name{gkrreg-package}
\alias{gkrreg}
\alias{gkrreg-package}
\title{gkrreg: Gaussian Kernel Robust Regression (GKRReg)}
\description{
Implements the Gaussian Kernel Robust Regression (GKRReg / GKRR) method proposed by De Carvalho, Lima Neto and Ferreira (2017) \doi{10.1016/j.neucom.2016.12.035}. The method re-weights observations iteratively using the Gaussian kernel so that poorly-fitted observations (outliers, leverage points) receive small weights, yielding resistance to Y-space outliers, X-space outliers and leverage points. Convergence is guaranteed by Propositions 4.1 and 4.2 of the original paper. Three estimators for the kernel width hyper-parameter are provided (S1: Caputo, S2: pairwise median, S3: residual variance). Inference is provided via an analytic sandwich variance estimator (default) or via bootstrap (percentile, normal and BCa intervals with p-values) through gkrr_boot(). Six real datasets from the robust regression literature are included to facilitate reproducible comparisons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcelorpf/gkrreg}
  \item Report bugs at \url{https://github.com/marcelorpf/gkrreg/issues}
}

}
\author{
\strong{Maintainer}: Marcelo Rodrigo Portela Ferreira \email{marcelo@de.ufpb.br}

Authors:
\itemize{
  \item Marcelo Rodrigo Portela Ferreira \email{marcelo@de.ufpb.br}
  \item Eufrásio de Andrade Lima Neto \email{eufrasio@de.ufpb.br}
}

}
\keyword{internal}
