% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kootenay}
\alias{kootenay}
\title{Kootenay River Water-Flow Measurements}
\format{
A data frame with 13 rows and 2 columns, with row names giving the
  year of measurement (1931--1943):
\describe{
  \item{libby}{Water flow at Libby, Montana (\eqn{10^6} cubic feet).}
  \item{newgate}{Water flow at Newgate, British Columbia
    (\eqn{10^6} cubic feet).}
}
}
\source{
J. Neter, M.H. Kutner, C.J. Nachtsheim, and W. Wasserman (1996).
\emph{Applied Linear Statistical Models}, 4th edition.
Irwin/McGraw-Hill, p. 414.

Available in \pkg{robustbase} as \code{kootenay}.
}
\usage{
kootenay
}
\description{
Annual water-flow measurements (in units of \eqn{10^6} cubic feet) at two
gauging stations on the Kootenay River: Libby (Montana, USA) and Newgate
(British Columbia, Canada).  The dataset contains a single X-space outlier
(year 1934) where the Libby measurement is anomalously large, likely due to
an upstream dam operation.
}
\details{
This dataset is used in De Carvalho et al. (2017), Section 6.5, to
illustrate robustness to X-space outliers.  The recommended width estimator
is \code{sigma_method = "s1"}.
}
\section{Outlier structure}{

The observation for year 1934 has an anomalously large \code{libby} value
(77.6 vs. a typical range of 17--39), making it an X-space outlier that
strongly influences naive regression fits.
}

\examples{
data(kootenay)

fit_ols  <- lm(newgate ~ libby, data = kootenay)
fit_gkrr <- gkrr(newgate ~ libby, data = kootenay, sigma_method = "s1")

plot(kootenay$libby, kootenay$newgate,
     xlab = "Libby flow (10^6 ft^3)", ylab = "Newgate flow (10^6 ft^3)",
     main = "Kootenay River Water Flow",
     pch = 19, col = "grey60")
# Highlight the X-space outlier (1934)
points(kootenay["1934", "libby"], kootenay["1934", "newgate"],
       col = "red", pch = 17, cex = 1.5)
abline(fit_ols,  col = "black",     lwd = 2, lty = 2)
abline(fit_gkrr, col = "firebrick", lwd = 2)
legend("topleft", c("OLS", "GKRR", "1934 (outlier)"),
       col = c("black","firebrick","red"),
       lty = c(2,1,NA), pch = c(NA,NA,17), lwd = 2, bty = "n")
}
\references{
De Carvalho, F.A.T., Lima Neto, E.A., Ferreira, M.R.P. (2017).
\doi{10.1016/j.neucom.2016.12.035}
}
\keyword{datasets}
