% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Brain and Body Mass of 62 Mammal Species}
\format{
A data frame with 62 rows and 5 columns:
\describe{
  \item{species}{Common name of the mammal species (character).}
  \item{body_mass}{Body mass in kilograms.}
  \item{brain_mass}{Brain mass in grams.}
  \item{log_body}{Natural logarithm of body mass.}
  \item{log_brain}{Natural logarithm of brain mass.}
}
}
\source{
P.J. Rousseeuw and A.M. Leroy (1987).
\emph{Robust Regression and Outlier Detection}.
John Wiley & Sons.

Originally from: T. Allison and D.V. Cicchetti (1976).
Sleep in mammals: ecological and constitutional correlates.
\emph{Science}, 194, 732--734.

Available in \pkg{MASS} as \code{mammals}.
}
\usage{
mammals
}
\description{
Body mass (kg) and brain mass (g) for 62 species of land mammals.
On the natural logarithm scale the relationship is approximately linear, but
the dataset contains several leverage points (very large mammals such as
African and Asian elephants) that strongly influence OLS estimates.
}
\section{Outlier structure}{

On the log scale, African elephant (\code{log_body} \eqn{\approx} 8.8) and
Asian elephant (\code{log_body} \eqn{\approx} 7.8) are high-leverage
observations.  Additionally, some primates (e.g., humans) deviate from the
overall trend (Y-space outliers relative to the bulk regression line).
}

\examples{
data(mammals)

fit_ols  <- lm(log_brain ~ log_body, data = mammals)
fit_gkrr <- gkrr(log_brain ~ log_body, data = mammals,
                 sigma_method = "s3")

plot(mammals$log_body, mammals$log_brain,
     xlab = "log(body mass, kg)", ylab = "log(brain mass, g)",
     main = "Brain vs. Body Mass (62 Mammal Species)",
     pch = 19, col = "grey60", cex = 0.8)
# Label the leverage points
elephants <- mammals$species \%in\% c("African elephant","Asian elephant")
points(mammals$log_body[elephants], mammals$log_brain[elephants],
       col = "red", pch = 17, cex = 1.4)
abline(fit_ols,  col = "black",     lwd = 2, lty = 2)
abline(fit_gkrr, col = "firebrick", lwd = 2)
legend("topleft", c("OLS","GKRR","Elephants"),
       col = c("black","firebrick","red"),
       lty = c(2,1,NA), pch = c(NA,NA,17), lwd = 2, bty = "n")
}
\keyword{datasets}
