% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stars_cyg}
\alias{stars_cyg}
\title{Hertzsprung-Russell Diagram of Star Cluster CYG OB1}
\format{
A data frame with 47 rows and 2 columns:
\describe{
  \item{log_temp}{Logarithm (base 10) of the effective surface temperature
    of the star (Kelvin).  Higher values correspond to hotter, bluer stars.}
  \item{log_light}{Logarithm (base 10) of the light intensity of the star
    relative to that of the Sun.}
}
}
\source{
Humphreys, R.M. (1978).
Studies of luminous stars in nearby galaxies.
\emph{Astrophysical Journal Supplement}, 38, 309--350.

Cited by: P.J. Rousseeuw and A.M. Leroy (1987).
\emph{Robust Regression and Outlier Detection}.
John Wiley & Sons, p. 27.

Available in \pkg{robustbase} as \code{starsCYG}.
}
\usage{
stars_cyg
}
\description{
The Hertzsprung-Russell diagram plots the log-luminosity against the
log-effective-temperature of 47 stars in the star cluster CYG OB1.
The dataset contains 4 giant stars (observations 11, 20, 30 and 34) that
are well off the main sequence and act as leverage points, severely
distorting OLS estimates of the linear trend in the main-sequence stars.
}
\section{Outlier structure}{

Observations 11, 20, 30 and 34 are giant stars that lie far from the main
sequence (low temperature, high luminosity).  They are leverage points
because their \code{log_temp} values are much lower than the bulk of the
data, and they also deviate from the linear relationship of the main sequence
(bad leverage points).
}

\examples{
data(stars_cyg)

fit_ols  <- lm(log_light ~ log_temp, data = stars_cyg)
fit_gkrr <- gkrr(log_light ~ log_temp, data = stars_cyg,
                 sigma_method = "s3")

plot(stars_cyg$log_temp, stars_cyg$log_light,
     xlab = "log10(Temperature, K)", ylab = "log10(Luminosity, Sun = 1)",
     main = "Hertzsprung-Russell Diagram (CYG OB1)",
     pch = 19, col = "grey60")
# Highlight the 4 giant stars
giants <- c(11, 20, 30, 34)
points(stars_cyg$log_temp[giants], stars_cyg$log_light[giants],
       col = "red", pch = 17, cex = 1.5)
abline(fit_ols,  col = "black",     lwd = 2, lty = 2)
abline(fit_gkrr, col = "firebrick", lwd = 2)
legend("topleft", c("OLS","GKRR","Giant stars"),
       col = c("black","firebrick","red"),
       lty = c(2,1,NA), pch = c(NA,NA,17), lwd = 2, bty = "n")
}
\references{
Rousseeuw, P.J. and Leroy, A.M. (1987).
\emph{Robust Regression and Outlier Detection}.
John Wiley & Sons.
}
\keyword{datasets}
